Move TreeComponent to a separate module
