+++ pyproject.toml
@@ -161,7 +161,7 @@ type = "virtualenv"
 "copy" = "gaphor.services.copyservice:CopyService"
 "sanitizer" = "gaphor.UML.sanitizerservice:SanitizerService"
 "namespace" = "gaphor.ui.namespace:Namespace"
-"tree_component" = "gaphor.ui.treemodel:TreeComponent"
+"tree_component" = "gaphor.ui.treecomponent:TreeComponent"
 "toolbox" = "gaphor.ui.toolbox:Toolbox"
 "diagrams" = "gaphor.ui.diagrams:Diagrams"
 "consolewindow" = "gaphor.plugins.console.consolewindow:ConsoleWindow"
